laboratory/requirements.txt
cryptography==45.0.2
python-dotenv==1.1.0
